requirements.txt
yfinance
pandas
numpy
beautifulsoup4
requests
vaderSentiment
streamlit
plotly
mysql-connector-python
scikit-learn
prophet
feedparser
joblib
lxml
python-dateutil